recipe: parse configuration in 'recipe' package
Recipe configuration file parsing has been moved from main package of 'debos'
tool to package 'recipe'. This change combines all things related to
'debos' configuration file in the single package allowing to re-use parsing
in other tools and simplify tests creation.
Package 'recipe' is now responsible for templating mechanism and YAML parsing.
