+++ recipe/recipe.go
@@ -1,9 +1,13 @@
 package recipe
 
 import (
+	"bytes"
+	"fmt"
 	"github.com/go-debos/debos"
 	"github.com/go-debos/debos/actions"
-	"log"
+	"gopkg.in/yaml.v2"
+	"path"
+	"text/template"
 )
 
 /* the YamlAction just embed the Action interface and implements the
@@ -52,10 +56,57 @@ func (y *YamlAction) UnmarshalYAML(unmarshal func(interface{}) error) error {
 	case "download":
 		y.Action = &actions.DownloadAction{}
 	default:
-		log.Fatalf("Unknown action: %v", aux.Action)
+		return fmt.Errorf("Unknown action: %v", aux.Action)
 	}
 
 	unmarshal(y.Action)
+
+	return nil
+}
+
+func sector(s int) int {
+	return s * 512
+}
+
+/*
+Parse method reads YAML recipe file and map all steps to appropriate actions.
+
+- file -- is the path to configuration file
+
+- templateVars -- optional argument allowing to use custom map for templating
+engine. Multiple template maps have no effect; only first map will be used.
+*/
+func (r *Recipe) Parse(file string, templateVars ...map[string]string) error {
+	t := template.New(path.Base(file))
+	funcs := template.FuncMap{
+		"sector": sector,
+	}
+	t.Funcs(funcs)
+
+	if _, err := t.ParseFiles(file); err != nil {
+		return err
+	}
+
+	if len(templateVars) == 0 {
+		templateVars = append(templateVars, make(map[string]string))
+	}
+
+	data := new(bytes.Buffer)
+	if err := t.Execute(data, templateVars[0]); err != nil {
+		return err
+	}
+
+	if err := yaml.Unmarshal(data.Bytes(), &r); err != nil {
+		return err
+	}
+
+	if len(r.Architecture) == 0 {
+		return fmt.Errorf("Recipe file must have 'architecture' property")
+	}
+
+	if len(r.Actions) == 0 {
+		return fmt.Errorf("Recipe file must have at least one action")
+	}
 
 	return nil
 }
